refactor(flotherm): split cross-platform helpers into lib/ subpackage
Move FloSCRIPT/FloXML/.pack/floerror.log code into
src/sim/drivers/flotherm/lib/ so it can be unit-tested on macOS / Linux
without Flotherm or pywinauto installed. The GUI driver and Win32 backend
stay where they are.

Architectural intent: lib/ is the future sim-flotherm PyPI package. We
defer extraction until lib/ exceeds ~1500 LOC, a second consumer wants
the API, or public release adds clear value. Until then, keeping the
boundary internal sidesteps the IP-audit conversation around vendor XSDs.

Moved:
  pack_project_dir, pack_project_name, lint_pack       → lib/pack.py
  lint_floscript, build_solve_*, build_custom          → lib/floscript.py
  lint_floxml                                          → lib/floxml.py
  read_floerror_log                                    → lib/error_log.py

Added 11 macOS-runnable lib/ unit tests (now 33 pass / 5 skip-on-missing-XSD).
